TM-161: Fix issue with future start date
Previous to this, if I set the start date for the future, clicking on any of the other dates would throw an exception. Need to check "previous" dates, as in "start" and "target" are previous to "urgent". Not any dates that are before TODAY. That never made sense, and it has been broken for a while.